requirements.txt
python-dotenv
pylint
flake8
pyyaml
cyclonedx-python-lib
